pyproject.toml
[tool.poetry]
name = "airline-test"
version = "0.1.0"
description = "API check-in aerolinea"
authors = ["kurotom"]
readme = "README.md"
packages = [{include = "aerolinea"}]

[tool.poetry.dependencies]
python = "^3.8"
mysqlclient = "^2.1.1"
django = "4.0"
djangorestframework = "3.13"
drf-yasg = "^1.21.5"
python-decouple = "^3.8"
whitenoise = "^6.4.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
